+++ src/cosl/rules.py
@@ -138,6 +138,11 @@ from .types import (
     RuleType,
     SingleRuleFormat,
 )
+# rules.py — backward-compat re-exports (deprecated, remove in next major)
+from .common_rules import ( 
+    generic_alert_groups, # pyright: ignore[reportUnusedImport]
+    HOST_METRICS_MISSING_RULE_NAME  # pyright: ignore[reportUnusedImport]
+) 
 
 logger = logging.getLogger(__name__)
